Move event definitions
