+++ core/ez/events/__init__.py
@@ -1,18 +1,18 @@
 from utilities.event import Event
 
-from .http import HTTP
 from .settings import Settings
+from web.events import HTTP
 from web.app.events import App
 from modules.events import Modules
-from plugins.events import Plugins
-from ezjsx.events import TreeRenderer
+# from plugins.events import Plugins
+# from ezjsx.events import TreeRenderer
 
 __all__ = [
     "Event",
     "HTTP",
     "App",
     "Modules",
-    "Plugins",
+    # "Plugins",
     "Settings",
     "TreeRenderer"
 ]
